Move time parsing stuff out to the to module.
